Address most of the review feedback
