+++ helper/xor/xor.go
@@ -1,35 +1,16 @@
 package xor
 
 import (
-	"crypto/rand"
 	"encoding/base64"
 	"fmt"
 )
 
-func GenerateRandBytes(length int) ([]byte, error) {
-	if length < 0 {
-		return nil, fmt.Errorf("length must be >= 0")
-	}
-
-	buf := make([]byte, length)
-	if length == 0 {
-		return buf, nil
-	}
-
-	n, err := rand.Read(buf)
-	if err != nil {
-		return nil, err
-	}
-	if n != length {
-		return nil, fmt.Errorf("unable to read %d bytes; only read %d", length, n)
-	}
-
-	return buf, nil
-}
-
-func XORBuffers(a, b []byte) ([]byte, error) {
+// XORBytes takes two byte slices and XORs them together, returning the final
+// byte slice. It is an error to pass in two byte slices that do not have the
+// same length.
+func XORBytes(a, b []byte) ([]byte, error) {
 	if len(a) != len(b) {
-		return nil, fmt.Errorf("length of buffers is not equivalent: %d != %d", len(a), len(b))
+		return nil, fmt.Errorf("length of byte slices is not equivalent: %d != %d", len(a), len(b))
 	}
 
 	buf := make([]byte, len(a))
@@ -41,6 +22,9 @@ func XORBuffers(a, b []byte) ([]byte, error) {
 	return buf, nil
 }
 
+// XORBase64 takes two base64-encoded strings and XORs the decoded byte slices
+// together, returning the final byte slice. It is an error to pass in two
+// strings that do not have the same length to their base64-decoded byte slice.
 func XORBase64(a, b string) ([]byte, error) {
 	aBytes, err := base64.StdEncoding.DecodeString(a)
 	if err != nil {
@@ -58,14 +42,5 @@ func XORBase64(a, b string) ([]byte, error) {
 		return nil, fmt.Errorf("decoded second base64 value is nil or empty")
 	}
 
-	if len(aBytes) != len(bBytes) {
-		return nil, fmt.Errorf("decoded values are not same length: %d != %d", len(aBytes), len(bBytes))
-	}
-
-	buf := make([]byte, len(aBytes))
-	for i, _ := range aBytes {
-		buf[i] = aBytes[i] ^ bBytes[i]
-	}
-
-	return buf, nil
+	return XORBytes(aBytes, bBytes)
 }
